Expose the bug
Using stopEditingWhenGridLosesFocus breaks Cypress tests under the following conditions:
1) The browser is Chromium-based (not with Firefox, and I don't know about Safari, etc.).
2) Using --watch mode
3) After the initial run (it should work), there's a change that forces recompilation

The break is that editing a cell gets stopped too soon. So, trying to enter the value "foo" with Cypress's type() function will mean that "f" will be accepted, and then editing will end. And then "o" will be accepted. And then "o" will be accepted.
Similarly (though not shown in this example), a custom editor will not work, because the editing session will end immediately.

See https://github.com/cypress-io/cypress/issues/5394 for an earlier example of the same behavior that was subsequently fixed. Whether this repo demonstrates a regression or a corner case that wasn't recognized at the time is unclear.